pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "pillow>=12.0.0",
    "pypandoc>=1.15",
    "python-magic>=0.4.27",
]
